[TASK] Integrate distribution manual from EXT:impexp documentation
Reduce the redundancy of the documentation by removing the chapter
"Distribution" from the EXT:impexp manual and integrating it into
the chapter "Creating a new distribution" of the official TYPO3
documentation "TYPO3 Explained".

Also set a convenient link to the EXT:impexp manual.

Also removed the warning for TYPO3 < 8.7.10 and < 9.1.0, which can be considered deprecated.

Also removed the screenshot as it does not provide any additional insight compared to the EXT:impexp manual.